refactor: separate db/types.ts file
