+++ app/trpc/server.ts
@@ -2,7 +2,8 @@ import { tinyassert } from "@hiogawa/utils";
 import { sql } from "drizzle-orm";
 import { z } from "zod";
 import { E, T, db, findOne } from "../db/drizzle-client.server";
-import { Q, Z_PRACTICE_ACTION_TYPES } from "../db/models";
+import { Q } from "../db/models";
+import { Z_PRACTICE_ACTION_TYPES } from "../db/types";
 import { importDeckJson } from "../misc/seed-utils";
 import { PracticeSystem } from "../utils/practice-system";
 import { middlewares } from "./context";
